requirements.txt
PyQt5>=5.15.0
pyserial>=3.5
pymodbus>=2.5.3 